+++ hidden/Extensions/Notification.Name+Extension.swift
@@ -13,4 +13,5 @@ extension Notification.Name {
     
     static let prefsChanged = Notification.Name("prefsChanged")
     static let alwayHideToggle = Notification.Name("alwayHideToggle")
+    static let revealForArranging = Notification.Name("revealForArranging")
 }